refactor(supervisor): unify deep_agent_single.py into deep_agent.py with CAIPE_DISTRIBUTED_AGENTS flag
Merge the single-node and multi-node supervisor implementations into a
single PlatformEngineerDeepAgent class in deep_agent.py, controlled by
the CAIPE_DISTRIBUTED_AGENTS env var (true = remote A2A registry,
false = local MCP sidecars).

Key changes:
- Delete deep_agent_single.py; deep_agent.py is now the single source
- Add _build_graph_multi_node() async method (lazy, was sync in __init__)
- Add _build_graph_async() for single-node MCP loading (unchanged)
- Both paths share: MongoDB checkpointer, Redis store, fact extraction,
  skills catalog, structured response, user_email/context_id support
- Rename ENABLE_AGENT_REGISTRY → CAIPE_DISTRIBUTED_AGENTS for clarity
- agent.py (multi-node): lazy graph init via _ensure_graph_initialized()
  called at the top of stream() instead of blocking __init__
- fastapi/main.py: eager startup hook calls ensure_initialized()
- graph.py: asyncio.run(ensure_initialized()) for LangGraph Studio
- All callers updated; deep_agent_single imports redirect to deep_agent
