+++ ai_platform_engineering/multi_agents/platform_engineer/graph.py
@@ -1,3 +1,6 @@
+import asyncio
 from ai_platform_engineering.multi_agents.platform_engineer.supervisor_agent import AIPlatformEngineerMAS
 
-graph = AIPlatformEngineerMAS().get_graph()
+_mas = AIPlatformEngineerMAS()
+asyncio.run(_mas.ensure_initialized())
+graph = _mas.get_graph()
