+++ ai_platform_engineering/utils/prompt_config.py
@@ -388,7 +388,7 @@ def generate_platform_system_prompt(config: Dict[str, Any], agents: Dict[str, An
     logger.info(f"System Prompt Template: {yaml_template}")
     
     if yaml_template:
-        # Provide rag_instructions placeholder - RAG is handled separately in deep_agent_single.py
+        # Provide rag_instructions placeholder - RAG is handled separately in deep_agent.py
         # Use empty string as default since deep agent adds RAG instructions after template formatting
         # final_answer_instructions is empty for single-node (uses structured response by default)
         return yaml_template.format(
