+++ docker-compose.dev.yaml
@@ -68,7 +68,7 @@ services:
       - AGENT_CONNECTIVITY_ENABLE_BACKGROUND=${AGENT_CONNECTIVITY_ENABLE_BACKGROUND:-true}
       - SKIP_AGENT_CONNECTIVITY_CHECK=${SKIP_AGENT_CONNECTIVITY_CHECK:-false}
       # Single-node mode: agents run as local MCP tools, no remote A2A registry needed
-      - ENABLE_AGENT_REGISTRY=false
+      - CAIPE_DISTRIBUTED_AGENTS=false
       - ENABLE_ARTIFACT_STREAMING=true
       - USE_STRUCTURED_RESPONSE=${USE_STRUCTURED_RESPONSE:-true}
